+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * region: Re-defined region tagging based on unicode standard rules, used when importing books, see server/clic/region
 * region: Initial text between parts and chapters is now part of the previous chapter, instead of being in it's own chapter. This will reduce chapter counts for books broken up into parts. [BREAKING CHANGE]
 * text: Shows book content verbatim from corpora repo, allows copy and paste
+* text: Hide embedded suspensions, since we no longer mark them up
 
 ## 1.7-beta3 (2018-09-21)